Adjust Picture form guide to Rails5.1 generated view code
The Installation guides do not specify a rails version:
`gem install rails --no-document`

Therefore the new people will install rails 5.1 (as of today).
One of the view changes was the introduction of `form_with` instead of `form_for`:
https://github.com/rails/rails/issues/25197

I updated both the 'replace the following code' sections as well as the edit code to
follow the new helper standards. (using the variable `form` instead of `f` \o/)